Use Imports/Suggests instead of Depends
This moves all the dependencies from Depends and Imports/Suggests. All functions that
are not used for graphs are in Imports, and all graphing functions are part of the
Suggests.

As part of the PEcAn project we are using Maeswrap to call the MAESPA model. We are
currently moving towards creating docker containers for each of the models and
discovered that Maeswrap depends on rgl, which adds an additional 500Mb to the
docker image due to dependencies that need to be installed (Mesa, X11). Moving the
rgl (as well as lattice and geometry) to Suggests removes this dependency on rgl.

We are also trying to remove as many dependencies that need to be loaded into the
namespace (using Depends) and moving all dependencies to Imports. This requires the
function to be called explicitly, for example str_trim is called using
stringr::str_trim.

For more information about this see also: https://github.com/PecanProject/pecan/pull/1758